Fixed cast-warnings on windows.
git-svn-id: https://svn.opendtect.org/od/trunk@32253 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/General/pickset.cc
@@ -508,7 +508,7 @@ float Set::getXYArea() const
     if ( polygon.isSelfIntersecting() )
 	return mUdf(float);
 
-    float area = polygon.area();
+    float area = (float) polygon.area();
     if ( SI().xyInFeet() )
 	area *= (mFromFeetFactorF*mFromFeetFactorF);
